+++ src/gsply/__init__.py
@@ -66,7 +66,7 @@ from gsply.utils import (
 )
 from gsply.writer import compress_to_arrays, compress_to_bytes, plywrite
 
-__version__ = "0.2.10"
+__version__ = "0.2.11"
 __all__ = [
     "plyread",
     "GSData",